refactor: move Name to user directory
